refactor: organize codes into separate files
