+++ drivers/emby/driver.go
@@ -1,15 +1,12 @@
 package emby
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"fmt"
-	"io"
 	"net/http"
 	"net/url"
 	"path"
-	"regexp"
 	"strings"
 	"time"
 
@@ -28,48 +25,6 @@ type Emby struct {
 	token  string
 	userID string
 }
-
-type authReq struct {
-	Username string `json:"Username"`
-	Pw       string `json:"Pw"`
-}
-
-type authResp struct {
-	AccessToken string `json:"AccessToken"`
-	User        struct {
-		ID string `json:"Id"`
-	} `json:"User"`
-}
-
-type listResp struct {
-	Items            []embyItem `json:"Items"`
-	TotalRecordCount int        `json:"TotalRecordCount"`
-}
-
-type embyItem struct {
-	Name        string `json:"Name"`
-	ID          string `json:"Id"`
-	Type        string `json:"Type"`
-	Path        string `json:"Path"`
-	SeriesName  string `json:"SeriesName"`
-	IndexNumber int    `json:"IndexNumber"`
-	ParentIndex int    `json:"ParentIndexNumber"`
-	IsFolder    bool   `json:"IsFolder"`
-	Size        int64  `json:"Size"`
-	DateCreated string `json:"DateCreated"`
-}
-
-type itemDetailResp struct {
-	MediaSources []embyMediaSource `json:"MediaSources"`
-}
-
-type embyMediaSource struct {
-	ID                   string `json:"Id"`
-	Container            string `json:"Container"`
-	SupportsDirectStream bool   `json:"SupportsDirectStream"`
-}
-
-var episodeCodeRegexp = regexp.MustCompile(`(?i)\bS\d{1,2}E\d{1,2}\b`)
 
 func (d *Emby) Config() driver.Config {
 	return config
@@ -300,126 +255,6 @@ func (d *Emby) Link(ctx context.Context, file model.Obj, args model.LinkArgs) (*
 			"User-Agent": []string{base.UserAgent},
 		},
 	}, nil
-}
-
-func (d *Emby) login(ctx context.Context) error {
-	payload, err := json.Marshal(authReq{
-		Username: d.Username,
-		Pw:       d.Password,
-	})
-	if err != nil {
-		return err
-	}
-
-	endpoint := d.URL + "/Users/AuthenticateByName"
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, endpoint, bytes.NewReader(payload))
-	if err != nil {
-		return err
-	}
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("X-Emby-Authorization", `MediaBrowser Client="OpenList", Device="OpenList", DeviceId="openlist-emby", Version="1.0.0"`)
-
-	resp, err := d.client.Do(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		body, _ := io.ReadAll(resp.Body)
-		return fmt.Errorf("emby auth failed: status=%d body=%s", resp.StatusCode, strings.TrimSpace(string(body)))
-	}
-
-	var data authResp
-	if err := json.NewDecoder(resp.Body).Decode(&data); err != nil {
-		return err
-	}
-	if strings.TrimSpace(data.AccessToken) == "" || strings.TrimSpace(data.User.ID) == "" {
-		return fmt.Errorf("emby auth response missing access token or user id")
-	}
-
-	d.token = data.AccessToken
-	d.userID = data.User.ID
-	return nil
-}
-
-func (d *Emby) getItems(ctx context.Context, parentID string) (*listResp, error) {
-	u, err := url.Parse(d.URL + "/Users/" + d.userID + "/Items")
-	if err != nil {
-		return nil, err
-	}
-	q := u.Query()
-	q.Set("ParentId", parentID)
-	q.Set("Recursive", "false")
-	q.Set("Fields", "Path,Size,DateCreated,SeriesName,IndexNumber,ParentIndexNumber")
-	q.Set("api_key", d.token)
-	u.RawQuery = q.Encode()
-
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, u.String(), nil)
-	if err != nil {
-		return nil, err
-	}
-
-	resp, err := d.client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		body, _ := io.ReadAll(resp.Body)
-		return nil, fmt.Errorf("emby list failed: status=%d body=%s", resp.StatusCode, strings.TrimSpace(string(body)))
-	}
-
-	var data listResp
-	if err := json.NewDecoder(resp.Body).Decode(&data); err != nil {
-		return nil, err
-	}
-	return &data, nil
-}
-
-func (d *Emby) MakeDir(ctx context.Context, parentDir model.Obj, dirName string) (model.Obj, error) {
-	return nil, errs.NotImplement
-}
-
-func (d *Emby) Move(ctx context.Context, srcObj, dstDir model.Obj) (model.Obj, error) {
-	return nil, errs.NotImplement
-}
-
-func (d *Emby) Rename(ctx context.Context, srcObj model.Obj, newName string) (model.Obj, error) {
-	return nil, errs.NotImplement
-}
-
-func (d *Emby) Copy(ctx context.Context, srcObj, dstDir model.Obj) (model.Obj, error) {
-	return nil, errs.NotImplement
-}
-
-func (d *Emby) Remove(ctx context.Context, obj model.Obj) error {
-	return errs.NotImplement
-}
-
-func (d *Emby) Put(ctx context.Context, dstDir model.Obj, file model.FileStreamer, up driver.UpdateProgress) (model.Obj, error) {
-	return nil, errs.NotImplement
-}
-
-func (d *Emby) GetArchiveMeta(ctx context.Context, obj model.Obj, args model.ArchiveArgs) (model.ArchiveMeta, error) {
-	return nil, errs.NotImplement
-}
-
-func (d *Emby) ListArchive(ctx context.Context, obj model.Obj, args model.ArchiveInnerArgs) ([]model.Obj, error) {
-	return nil, errs.NotImplement
-}
-
-func (d *Emby) Extract(ctx context.Context, obj model.Obj, args model.ArchiveInnerArgs) (*model.Link, error) {
-	return nil, errs.NotImplement
-}
-
-func (d *Emby) ArchiveDecompress(ctx context.Context, srcObj, dstDir model.Obj, args model.ArchiveDecompressArgs) ([]model.Obj, error) {
-	return nil, errs.NotImplement
-}
-
-func (d *Emby) GetDetails(ctx context.Context) (*model.StorageDetails, error) {
-	return nil, errs.NotImplement
 }
 
 var _ driver.Driver = (*Emby)(nil)
